# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: GearBy API
  version: 1.0.0
  description: Executable API contract derived from the GearBy Source of Truth. P0 is the first MVP and recommendation endpoints describe the planned P1 boundary.
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /stores/{storeId}:
    get:
      operationId: getStore
      summary: Get one published store
      parameters: [{ $ref: '#/components/parameters/StoreId' }]
      responses:
        '200':
          $ref: '#/components/responses/ApiResponse'
        '404': { $ref: '#/components/responses/ApiError' }
...
  /admin/stores:
    get:
      operationId: listAdminStores
      summary: List stores across review states
      security: [{ adminOidc: [ADMIN] }]
      responses:
        '200':
          $ref: '#/components/responses/ApiResponse'
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
    post:
      operationId: createAdminStore
      summary: Create a draft store for review
      security: [{ adminOidc: [ADMIN] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreInput' }
      responses:
        '201':
          $ref: '#/components/responses/ApiResponse'
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
  /admin/stores/{storeId}:
    patch:
      operationId: updateAdminStore
      summary: Update a draft or review store
      security: [{ adminOidc: [ADMIN] }]
      parameters: [{ $ref: '#/components/parameters/StoreId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreInput' }
      responses:
        '200':
          $ref: '#/components/responses/ApiResponse'
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
  /admin/stores/{storeId}/publish:
    post:
      operationId: publishAdminStore
      summary: Publish a reviewed store or re-verify a published store
      description: Repeating this operation for a published store records a new administrator verification time.
      security: [{ adminOidc: [ADMIN] }]
      parameters: [{ $ref: '#/components/parameters/StoreId' }]
      responses:
        '200':
          $ref: '#/components/responses/ApiResponse'
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
  /admin/stores/{storeId}/review:
    post:
      operationId: reviewAdminStore
      summary: Submit a draft or rejected store for review
      security: [{ adminOidc: [ADMIN] }]
      parameters: [{ $ref: '#/components/parameters/StoreId' }]
      responses:
        '200':
          $ref: '#/components/responses/ApiResponse'
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
  /admin/stores/{storeId}/reject:
    post:
      operationId: rejectAdminStore
      summary: Reject a reviewed store back to an editable state
      security: [{ adminOidc: [ADMIN] }]
      parameters: [{ $ref: '#/components/parameters/StoreId' }]
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/StoreRejectionInput' } } }
      responses:
        '200':
          $ref: '#/components/responses/ApiResponse'
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
...
components:
  securitySchemes:
    adminOidc:
      type: oauth2
      description: Invite-only administrator authentication; public administrator signup is not provided.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.com/authorize
          tokenUrl: https://identity.example.com/token
          scopes: { ADMIN: GearBy administrator access }
  parameters:
    StoreId: { name: storeId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    ApiResponse:
      description: Successful response envelope. The operation-specific value is in `data`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    ApiError:
      description: Failed response envelope. The HTTP status and `error.code` describe the failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
...
  schemas:
    ApiResponse:
      type: object
      required: [success, timestamp, data, error]
      properties:
        success: { type: boolean }
        timestamp: { type: string, format: date-time }
        data: {}
        error:
          oneOf:
            - $ref: '#/components/schemas/ApiError'
            - type: 'null'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL_ERROR] }
        message: { type: string }
...
    CategorySlug: { type: string, enum: [HIKING, BACKPACKING, CAMPING, CLIMBING] }
...
    Coordinates:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
...
    StoreInput:
      type: object
      required: [name, address, coordinates, categories]
      properties:
        name: { type: string, minLength: 1 }
        address: { type: string, minLength: 1 }
        coordinates: { $ref: '#/components/schemas/Coordinates' }
        categories: { type: array, minItems: 1, items: { $ref: '#/components/schemas/CategorySlug' } }
        phone: { type: string }
        hours: { type: string }
        description: { type: string }
    StoreRejectionInput:
      type: object
      properties:
        reason: { type: string, minLength: 1, maxLength: 500 }
...